pyproject.toml
[project]
name = "lean4-remote-worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.3",
    "pexpect>=4.9.0",
    "pycryptodome>=3.23.0",
    "requests>=2.32.5",
]
